+++ src/bittytax/bt_types.py
@@ -6,6 +6,22 @@ from enum import Enum
 from typing import NewType
 
 
+class TaxRules(Enum):
+    UK_INDIVIDUAL = "UK Individual"
+    UK_COMPANY_JAN = "UK Company (Jan)"
+    UK_COMPANY_FEB = "UK Company (Feb)"
+    UK_COMPANY_MAR = "UK Company (Mar)"
+    UK_COMPANY_APR = "UK Company (Apr)"
+    UK_COMPANY_MAY = "UK Company (May)"
+    UK_COMPANY_JUN = "UK Company (Jun)"
+    UK_COMPANY_JUL = "UK Company (Jul)"
+    UK_COMPANY_AUG = "UK Company (Aug)"
+    UK_COMPANY_SEP = "UK Company (Sep)"
+    UK_COMPANY_OCT = "UK Company (Oct)"
+    UK_COMPANY_NOV = "UK Company (Nov)"
+    UK_COMPANY_DEC = "UK Company (Dec)"
+
+
 class TrType(Enum):
     DEPOSIT = "Deposit"
     MINING = "Mining"
@@ -21,6 +37,7 @@ class TrType(Enum):
     FEE_REBATE = "Fee-Rebate"
     LOAN = "Loan"
     MARGIN_GAIN = "Margin-Gain"
+    MARGIN_FEE_REBATE = "Margin-Fee-Rebate"
     WITHDRAWAL = "Withdrawal"
     SPEND = "Spend"
     GIFT_SENT = "Gift-Sent"
@@ -49,6 +66,21 @@ class TrRecordPart(Enum):
     FEE = "Fee"
 
 
+TAX_RULES_UK_COMPANY = [
+    TaxRules.UK_COMPANY_JAN,
+    TaxRules.UK_COMPANY_FEB,
+    TaxRules.UK_COMPANY_MAR,
+    TaxRules.UK_COMPANY_APR,
+    TaxRules.UK_COMPANY_MAY,
+    TaxRules.UK_COMPANY_JUN,
+    TaxRules.UK_COMPANY_JUL,
+    TaxRules.UK_COMPANY_AUG,
+    TaxRules.UK_COMPANY_SEP,
+    TaxRules.UK_COMPANY_OCT,
+    TaxRules.UK_COMPANY_NOV,
+    TaxRules.UK_COMPANY_DEC,
+]
+
 BUY_TYPES = (
     TrType.DEPOSIT,
     TrType.MINING,
@@ -64,6 +96,7 @@ BUY_TYPES = (
     TrType.FEE_REBATE,
     TrType.LOAN,
     TrType.MARGIN_GAIN,
+    TrType.MARGIN_FEE_REBATE,
     TrType.TRADE,
 )