+++ src/bittytax/report.py
@@ -18,14 +18,15 @@ from colorama import Fore, Style
 from xhtml2pdf import pisa
 
 from .audit import AuditRecords, AuditTotals
-from .bt_types import AssetName, AssetSymbol, Date, Note, Year
+from .bt_types import TAX_RULES_UK_COMPANY, AssetName, AssetSymbol, Date, Note, TaxRules, Year
 from .config import config
-from .constants import _H1, ERROR, H1, TAX_RULES_UK_COMPANY
+from .constants import _H1, ERROR, H1
 from .price.valueasset import VaPriceReport
 from .tax import (
     CalculateCapitalGains,
     CalculateIncome,
     CalculateMarginTrading,
+    CapitalGainsReportEstimate,
     HoldingsReportRecord,
     TaxReportRecord,
 )
@@ -232,7 +233,7 @@ class ReportLog:
 
             self._tax_full(args.tax_rules, audit, tax_report, price_report, holdings_report)
 
-    def _tax_summary(self, tax_rules: str, tax_report: Dict[Year, TaxReportRecord]) -> None:
+    def _tax_summary(self, tax_rules: TaxRules, tax_report: Dict[Year, TaxReportRecord]) -> None:
         print(f"{Fore.WHITE}tax report output:")
         for tax_year in sorted(tax_report):
             print(
@@ -249,7 +250,7 @@ class ReportLog:
 
     def _tax_full(
         self,
-        tax_rules: str,
+        tax_rules: TaxRules,
         audit: AuditRecords,
         tax_report: Dict[Year, TaxReportRecord],
         price_report: Dict[Year, Dict[AssetSymbol, Dict[Date, VaPriceReport]]],
@@ -361,7 +362,7 @@ class ReportLog:
                     f'{"":<28} {self.format_quantity(quantity):>25} '
                     f"{self.format_value(cost):>13} {self.format_value(fees):>13} "
                     f"{self.format_value(proceeds):>13} "
-                    f"{Fore.RED if gain < 0 else Fore.WHITE}{self.format_value(gain):>13}"
+                    f"{Fore.RED if gain < 0 else Fore.YELLOW}{self.format_value(gain):>13}"
                 )
 
         print(f'{Fore.YELLOW}{"_" * len(header)}')
@@ -376,7 +377,7 @@ class ReportLog:
 
         print(f"\n{Fore.CYAN}Summary\n")
         print(f'{Fore.WHITE}{"Number of disposals:":<40} {cgains.summary["disposals"]:>13}')
-        if cgains.cgt_estimate["proceeds_warning"]:
+        if cgains.summary["proceeds_warning"]:
             print(
                 f'{Fore.WHITE}{"Disposal proceeds:":<40} '
                 f'{"*" + self.format_value(cgains.totals["proceeds"]):>13}'.replace(
@@ -404,10 +405,23 @@ class ReportLog:
             f'{self.format_value(abs(cgains.summary["total_loss"])):>13}'
         )
 
-        if cgains.cgt_estimate["proceeds_warning"]:
+        if cgains.split_date:
+            print(
+                f'\n{Fore.WHITE}{f"Gains before {self.format_date2(cgains.split_date)}":<40} '
+                f'{Fore.RED if cgains.split_totals["gain_before"] < 0 else Fore.WHITE}'
+                f'{self.format_value(cgains.split_totals["gain_before"]):>13}'
+            )
+
+            print(
+                f'{Fore.WHITE}{f"Gains on or after {self.format_date2(cgains.split_date)}":<40} '
+                f'{Fore.RED if cgains.split_totals["gain_after"] < 0 else Fore.WHITE}'
+                f'{self.format_value(cgains.split_totals["gain_after"]):>13}'
+            )
+
+        if cgains.summary["proceeds_limit"] is not None and cgains.summary["proceeds_warning"]:
             print(
                 f"{Fore.YELLOW}*Assets sold are more than "
-                f'{self.format_value(cgains.cgt_estimate["proceeds_limit"])}, '
+                f'{self.format_value(cgains.summary["proceeds_limit"])}, '
                 "this needs to be reported to HMRC if you already complete a Self Assessment"
             )
 
@@ -416,37 +430,48 @@ class ReportLog:
         print(
             f"{Fore.CYAN}The figures below are only an estimate, "
             "they do not take into consideration other gains and losses in the same tax year, "
-            "always consult with a professional accountant before filing.\n"
+            "always consult with a professional accountant before filing."
         )
-        if cgains.totals["gain"] > 0:
+
+        if not cgains.split_date:
+            self._cgt_estimate2(cgains.cgt_estimate)
+        else:
+            print(f"\n{Fore.CYAN}Gains before {self.format_date2(cgains.split_date)}")
+            self._cgt_estimate2(cgains.cgt_estimate)
+            print(f"\n{Fore.CYAN}Gains on or after {self.format_date2(cgains.split_date)}")
+            self._cgt_estimate2(cgains.cgt_estimate_split)
+
+    def _cgt_estimate2(self, cgt_estimate: CapitalGainsReportEstimate) -> None:
+        if cgt_estimate["allowance_used"]:
             print(
-                f'{Fore.WHITE}{"Taxable Gain*:":<40} '
-                f'{self.format_value(cgains.cgt_estimate["taxable_gain"]):>13}'.replace(
+                f'\n{Fore.WHITE}{"Taxable Gain*:":<40} '
+                f'{self.format_value(cgt_estimate["taxable_gain"]):>13}'.replace(
                     "*", f"{Fore.YELLOW}*{Fore.WHITE}"
                 )
             )
         else:
             print(
-                f'{Fore.WHITE}{"Taxable Gain:":<40} '
-                f'{self.format_value(cgains.cgt_estimate["taxable_gain"]):>13}'
+                f'\n{Fore.WHITE}{"Taxable Gain:":<40} '
+                f'{self.format_value(cgt_estimate["taxable_gain"]):>13}'
             )
 
         print(
             f'{Fore.WHITE}{"Capital Gains Tax (Basic rate):":<40} '
-            f'{self.format_value(cgains.cgt_estimate["cgt_basic"]):>13} '
-            f'({self.format_rate(cgains.cgt_estimate["cgt_basic_rate"])})'
+            f'{self.format_value(cgt_estimate["cgt_basic"]):>13} '
+            f'({self.format_rate(cgt_estimate["cgt_basic_rate"])})'
         )
 
         print(
             f'{Fore.WHITE}{"Capital Gains Tax (Higher rate):":<40} '
-            f'{self.format_value(cgains.cgt_estimate["cgt_higher"]):>13} '
-            f'({self.format_rate(cgains.cgt_estimate["cgt_higher_rate"])})'
+            f'{self.format_value(cgt_estimate["cgt_higher"]):>13} '
+            f'({self.format_rate(cgt_estimate["cgt_higher_rate"])})'
         )
 
-        if cgains.cgt_estimate["allowance_used"]:
+        if cgt_estimate["allowance_used"]:
             print(
-                f'{Fore.YELLOW}*{self.format_value(cgains.cgt_estimate["allowance_used"])} of the '
-                f'tax-free allowance ({self.format_value(cgains.cgt_estimate["allowance"])}) used'
+                f'{Fore.YELLOW}*{self.format_value(cgt_estimate["allowance_used"])} of the '
+                f'"Annual Exempt Amount" ({self.format_value(cgt_estimate["allowance"])}) '
+                f"tax-free allowance used"
             )
 
     def _ct_estimate(self, cgains: CalculateCapitalGains) -> None:
@@ -480,8 +505,8 @@ class ReportLog:
     def _income(self, income: CalculateIncome) -> None:
         print(f"\n{Fore.CYAN}Income\n")
         header = (
-            f'{"Asset":<{self.MAX_SYMBOL_LEN}} {"Date":<10} {"Type":<10} {"Description":<40} '
-            f'{"Quantity":<25} {"Amount":>13} {"Fees":>13}'
+            f'{"Asset":<{self.MAX_SYMBOL_LEN}} {"Date":<10} {"Income Type":<15} '
+            f'{"Description":<40} {"Quantity":>25} {"Market Value":>15} {"Fees":>13}'
         )
 
         print(f"{Fore.YELLOW}{header}")
@@ -494,40 +519,43 @@ class ReportLog:
                 fees += te.fees
                 print(
                     f"{Fore.WHITE}{te.asset:<{self.MAX_SYMBOL_LEN}} "
-                    f"{self.format_date(te.date):<10} {te.type.value:<10} "
-                    f"{self.format_note(te.note):<40} {self.format_quantity(te.quantity):<25} "
-                    f"{self.format_value(te.amount):>13} {self.format_value(te.fees):>13}"
+                    f"{self.format_date(te.date):<10} {te.type.value:<15} "
+                    f"{self.format_note(te.note):<40} {self.format_quantity(te.quantity):>25} "
+                    f"{self.format_value(te.amount):>15} {self.format_value(te.fees):>13}"
                 )
             if events > 1:
                 print(
-                    f'{Fore.YELLOW}{"Total":<{self.MAX_SYMBOL_LEN}} {"":<10} '
-                    f'{"":<10} {"":<40} {self.format_quantity(quantity):<25} '
-                    f"{self.format_value(amount):>13} {self.format_value(fees):>13}\n"
+                    f'{Fore.YELLOW}{"Total":<{self.MAX_SYMBOL_LEN}} {"":<15} '
+                    f'{"":<10} {"":<40} {self.format_quantity(quantity):>25} '
+                    f"{self.format_value(amount):>15} {self.format_value(fees):>13}\n"
                 )
 
         print(
-            f'{Fore.YELLOW}{"Income Type":<{self.MAX_SYMBOL_LEN + 11}} {"":<10} '
-            f'{"":<40} {"":<25} {"Amount":>13} {"Fees":>13}'
+            f'{Fore.YELLOW}{"Income Type":<{self.MAX_SYMBOL_LEN + 11}} {"":<15} '
+            f'{"":<40} {"":>25} {"Market Value":>15} {"Fees":>13}'
         )
 
         for i_type in sorted(income.type_totals):
             print(
-                f'{Fore.WHITE}{i_type:<{self.MAX_SYMBOL_LEN + 11}} {"":<10} '
-                f'{"":<40} {"":<25} {self.format_value(income.type_totals[i_type]["amount"]):>13} '
+                f'{Fore.WHITE}{i_type:<{self.MAX_SYMBOL_LEN + 11}} {"":<15} '
+                f'{"":<40} {"":>25} {self.format_value(income.type_totals[i_type]["amount"]):>15} '
                 f'{self.format_value(income.type_totals[i_type]["fees"]):>13}'
             )
 
         print(f'{Fore.YELLOW}{"_" * len(header)}')
         print(
-            f'{Fore.YELLOW}{Style.BRIGHT}{"Total":<{self.MAX_SYMBOL_LEN + 11}} {"":<10} '
-            f'{"":<40} {"":<25} {self.format_value(income.totals["amount"]):>13} '
+            f'{Fore.YELLOW}{Style.BRIGHT}{"Total":<{self.MAX_SYMBOL_LEN + 11}} {"":<15} '
+            f'{"":<40} {"":>25} {self.format_value(income.totals["amount"]):>15} '
             f'{self.format_value(income.totals["fees"]):>13}{Style.NORMAL}'
         )
 
     def _margin_trading(self, margin: CalculateMarginTrading) -> None:
         print(f"\n{Fore.CYAN}Margin Trading\n")
         print("These figures are NOT included in the Summary or Tax Estimate above.\n")
-        header = f'{"Wallet":<30} {"Contract":<40} {"Gains":>13} {"Losses":>13} {"Fees":>13}'
+        header = (
+            f'{"Wallet":<30} {"Contract":<40} {"Gains":>13} {"Losses":>13} {"Fees":>13} '
+            f'{"Fee Rebates":>13}'
+        )
 
         print(f"{Fore.YELLOW}{header}")
         for wallet, note in sorted(
@@ -538,7 +566,8 @@ class ReportLog:
                 f"{Fore.WHITE}{self.format_note(note):<40} "
                 f'{self.format_value(margin.contract_totals[(wallet, note)]["gains"]):>13} '
                 f'{self.format_value(margin.contract_totals[(wallet, note)]["losses"]):>13} '
-                f'{self.format_value(margin.contract_totals[(wallet, note)]["fees"]):>13}'
+                f'{self.format_value(margin.contract_totals[(wallet, note)]["fees"]):>13} '
+                f'{self.format_value(margin.contract_totals[(wallet, note)]["fee_rebates"]):>13}'
                 f"{Style.NORMAL}"
             )
 
@@ -548,7 +577,8 @@ class ReportLog:
             f'{"":<40} '
             f'{self.format_value(margin.totals["gains"]):>13} '
             f'{self.format_value(margin.totals["losses"]):>13} '
-            f'{self.format_value(margin.totals["fees"]):>13}{Style.NORMAL}'
+            f'{self.format_value(margin.totals["fees"]):>13} '
+            f'{self.format_value(margin.totals["fee_rebates"]):>13}{Style.NORMAL}'
         )
 
     def _price_data(self, price_report: Dict[AssetSymbol, Dict[Date, VaPriceReport]]) -> None:
@@ -581,8 +611,8 @@ class ReportLog:
         print(f"{Fore.CYAN}Current Holdings\n")
 
         header = (
-            f'{"Asset":<{self.ASSET_WIDTH}} {"Quantity":>25} {"Cost + Fees":>16} {"Value":>16} '
-            f'{"Gain":>16}'
+            f'{"Asset":<{self.ASSET_WIDTH}} {"Quantity":>25} {"Cost + Fees":>16} '
+            f'{"Market Value":>16} {"Gain":>16}'
         )
 
         print(f"{Fore.YELLOW}{header}")